requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autoflake==2.3.1
billiard==4.2.0
black==24.3.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.1
colorama==0.4.6
coverage==7.4.4
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.3
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-data-browser==4.2.8
django-debug-toolbar==4.3.0
django-etc==1.4.0
django-extensions==3.2.3
django-filter==24.2
django-hitcount==1.3.5
django-jazzmin==2.6.1
django-lifecycle==1.2.3
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
drf-spectacular==0.27.1
exceptiongroup==1.2.0
execnet==2.0.2
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
hyperlink==21.0.0
identify==2.5.35
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
loguru==0.7.2
mailjet-rest==1.3.4
model-bakery==1.17.0
msgpack==1.0.8
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.33
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
pyrsistent==0.20.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0
pytest-xdist==3.5.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
ruff==0.3.4
sentry-sdk==1.44.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
whitenoise==6.6.0
